Move methods from Feed ControllerProvider to Controller
