+++ lib/Alchemy/Phrasea/ControllerProvider/Prod/Feed.php
@@ -11,29 +11,32 @@
 
 namespace Alchemy\Phrasea\ControllerProvider\Prod;
 
-use Alchemy\Phrasea\Controller\RecordsRequest;
+use Alchemy\Phrasea\Application as PhraseaApplication;
+use Alchemy\Phrasea\Controller\Prod\FeedController;
 use Alchemy\Phrasea\ControllerProvider\ControllerProviderTrait;
-use Alchemy\Phrasea\Core\Event\FeedEntryEvent;
-use Alchemy\Phrasea\Core\PhraseaEvents;
-use Alchemy\Phrasea\Feed\Aggregate;
-use Alchemy\Phrasea\Feed\Link\AggregateLinkGenerator;
-use Alchemy\Phrasea\Feed\Link\FeedLinkGenerator;
-use Alchemy\Phrasea\Model\Entities\FeedEntry;
-use Alchemy\Phrasea\Model\Entities\FeedItem;
 use Silex\Application;
 use Silex\ControllerProviderInterface;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
+use Silex\ServiceProviderInterface;
 
-class Feed implements ControllerProviderInterface
+class Feed implements ControllerProviderInterface, ServiceProviderInterface
 {
     use ControllerProviderTrait;
 
+    public function register(Application $app)
+    {
+        $app['controller.prod.feed'] = $app->share(function (PhraseaApplication $app) {
+            return (new FeedController($app))
+                ->setFirewall($app['firewall']);
+        });
+    }
+
+    public function boot(Application $app)
+    {
+        // no-op
+    }
+
     public function connect(Application $app)
     {
-        $app['controller.prod.feed'] = $this;
-
         $controllers = $this->createAuthenticatedCollection($app);
 
         $controllers->post('/requestavailable/', 'controller.prod.feed:publishRecordsAction');
@@ -72,272 +75,5 @@ class Feed implements ControllerProviderInterface
             ->assert('id', '\d+');
 
         return $controllers;
-    }
-
-    public function publishRecordsAction(Application $app, Request $request)
-    {
-        $feeds = $app['repo.feeds']->getAllForUser(
-            $app['acl']->get($app['authentication']->getUser())
-        );
-        $publishing = RecordsRequest::fromRequest($app, $request, true, [], ['bas_chupub']);
-
-        return $app['twig']->render(
-            'prod/actions/publish/publish.html.twig',
-            ['publishing' => $publishing, 'feeds' => $feeds]
-        );
-    }
-
-    public function createFeedEntryAction(Application $app, Request $request) {
-        $feed = $app['repo.feeds']->find($request->request->get('feed_id'));
-
-        if (null === $feed) {
-            $app->abort(404, "Feed not found");
-        }
-
-        $publisher = $app['repo.feed-publishers']->findOneBy(
-            ['feed' => $feed, 'user' => $app['authentication']->getUser()]
-        );
-
-        if ('' === $title = trim($request->request->get('title', ''))) {
-            $app->abort(400, "Bad request");
-        }
-
-        if (!$feed->isPublisher($app['authentication']->getUser())) {
-            $app->abort(403, 'Unathorized action');
-        }
-
-        $entry = new FeedEntry();
-        $entry->setAuthorEmail($request->request->get('author_mail'))
-            ->setAuthorName($request->request->get('author_name'))
-            ->setTitle($title)
-            ->setFeed($feed)
-            ->setPublisher($publisher)
-            ->setSubtitle($request->request->get('subtitle', ''));
-
-        $feed->addEntry($entry);
-
-        $publishing = RecordsRequest::fromRequest($app, $request, true, [], ['bas_chupub']);
-        foreach ($publishing as $record) {
-            $item = new FeedItem();
-            $item->setEntry($entry)
-                ->setRecordId($record->get_record_id())
-                ->setSbasId($record->get_sbas_id());
-            $entry->addItem($item);
-            $app['orm.em']->persist($item);
-        }
-
-        $app['orm.em']->persist($entry);
-        $app['orm.em']->persist($feed);
-        $app['orm.em']->flush();
-
-        $app['dispatcher']->dispatch(
-            PhraseaEvents::FEED_ENTRY_CREATE,
-            new FeedEntryEvent($entry, $request->request->get('notify'))
-        );
-
-        $datas = ['error' => false, 'message' => false];
-
-        return $app->json($datas);
-    }
-
-    public function editEntryAction(Application $app, Request $request, $id) {
-        $entry = $app['repo.feed-entries']->find($id);
-
-        if (!$entry->isPublisher($app['authentication']->getUser())) {
-            throw new AccessDeniedHttpException();
-        }
-
-        $feeds = $app['repo.feeds']->getAllForUser($app['acl']->get($app['authentication']->getUser()));
-
-        $datas = $app['twig']->render(
-            'prod/actions/publish/publish_edit.html.twig',
-            ['entry' => $entry, 'feeds' => $feeds]
-        );
-
-        return new Response($datas);
-    }
-
-    /**
-     * @param Application $app
-     * @param Request     $request
-     * @param int         $id
-     * @return \Symfony\Component\HttpFoundation\JsonResponse
-     */
-    public function updateEntryAction(Application $app, Request $request, $id) {
-        $datas = ['error' => true, 'message' => '', 'datas' => ''];
-        $entry = $app['repo.feed-entries']->find($id);
-
-        if (null === $entry) {
-            $app->abort(404, 'Entry not found');
-        }
-        if (!$entry->isPublisher($app['authentication']->getUser())) {
-            $app->abort(403, 'Unathorized action');
-        }
-        if ('' === $title = trim($request->request->get('title', ''))) {
-            $app->abort(400, "Bad request");
-        }
-
-        $entry->setAuthorEmail($request->request->get('author_mail'))
-            ->setAuthorName($request->request->get('author_name'))
-            ->setTitle($title)
-            ->setSubtitle($request->request->get('subtitle', ''));
-
-        $currentFeedId = $entry->getFeed()
-            ->getId();
-        $new_feed_id = $request->request->get('feed_id', $currentFeedId);
-        if ($currentFeedId !== (int)$new_feed_id) {
-
-            $new_feed = $app['repo.feeds']->find($new_feed_id);
-
-            if ($new_feed === null) {
-                $app->abort(404, 'Feed not found');
-            }
-
-            if (!$new_feed->isPublisher($app['authentication']->getUser())) {
-                $app->abort(403, 'You are not publisher of this feed');
-            }
-            $entry->setFeed($new_feed);
-        }
-
-        $items = explode(';', $request->request->get('sorted_lst'));
-
-        foreach ($items as $item_sort) {
-            $item_sort_datas = explode('_', $item_sort);
-            if (count($item_sort_datas) != 2) {
-                continue;
-            }
-            $item = $app['repo.feed-items']->find($item_sort_datas[0]);
-            $item->setOrd($item_sort_datas[1]);
-            $app['orm.em']->persist($item);
-        }
-
-        $app['orm.em']->persist($entry);
-        $app['orm.em']->flush();
-
-        return $app->json(
-            [
-                'error'   => false,
-                'message' => 'succes',
-                'datas'   => $app['twig']->render(
-                    'prod/results/entry.html.twig',
-                    [
-                        'entry' => $entry
-                    ]
-                )
-            ]
-        );
-    }
-
-    public function deleteEntryAction(Application $app, Request $request, $id) {
-        $datas = ['error' => true, 'message' => ''];
-
-        $entry = $app['repo.feed-entries']->find($id);
-
-        if (null === $entry) {
-            $app->abort(404, 'Entry not found');
-        }
-        if (!$entry->isPublisher($app['authentication']->getUser()) && $entry->getFeed()
-                ->isOwner($app['authentication']->getUser()) === false
-        ) {
-            $app->abort(403, $app->trans('Action Forbidden : You are not the publisher'));
-        }
-
-        $app['orm.em']->remove($entry);
-        $app['orm.em']->flush();
-
-        return $app->json(['error' => false, 'message' => 'succes']);
-    }
-
-    public function indexAction(Application $app, Request $request) {
-        $request = $app['request'];
-        $page = (int)$request->query->get('page');
-        $page = $page > 0 ? $page : 1;
-
-        $feeds = $app['repo.feeds']->getAllForUser($app['acl']->get($app['authentication']->getUser()));
-
-        $datas = $app['twig']->render(
-            'prod/results/feeds.html.twig',
-            [
-                'feeds' => $feeds,
-                'feed'  => new Aggregate($app['orm.em'], $feeds),
-                'page'  => $page
-            ]
-        );
-
-        return new Response($datas);
-    }
-
-    public function showAction(Application $app, Request $request, $id) {
-        $page = (int)$request->query->get('page');
-        $page = $page > 0 ? $page : 1;
-
-        $feed = $app['repo.feeds']->find($id);
-        if (!$feed->isAccessible($app['authentication']->getUser(), $app)) {
-            $app->abort(404, 'Feed not found');
-        }
-        $feeds = $app['repo.feeds']->getAllForUser($app['acl']->get($app['authentication']->getUser()));
-
-        $datas = $app['twig']->render(
-            'prod/results/feeds.html.twig',
-            ['feed' => $feed, 'feeds' => $feeds, 'page' => $page]
-        );
-
-        return new Response($datas);
-    }
-
-    public function subscribeAggregatedFeedAction(Application $app, Request $request) {
-        $renew = ($request->query->get('renew') === 'true');
-
-        $feeds = $app['repo.feeds']->getAllForUser($app['acl']->get($app['authentication']->getUser()));
-
-        $link = $app['feed.aggregate-link-generator']->generate(
-            new Aggregate($app['orm.em'], $feeds),
-            $app['authentication']->getUser(),
-            AggregateLinkGenerator::FORMAT_RSS,
-            null,
-            $renew
-        );
-
-        $output = [
-            'texte' => '<p>' . $app->trans(
-                    'publication::Voici votre fil RSS personnel. Il vous permettra d\'etre tenu au courrant des publications.'
-                ) . '</p><p>' . $app->trans('publications::Ne le partagez pas, il est strictement confidentiel') . '</p>
-            <div><input type="text" readonly="readonly" class="input_select_copy" value="' . $link->getURI()
-                . '"/></div>',
-            'titre' => $app->trans('publications::votre rss personnel')
-        ];
-
-        return $app->json($output);
-    }
-
-    public function subscribeFeedAction(Application $app, Request $request, $id) {
-        $renew = ($request->query->get('renew') === 'true');
-
-        $feed = $app['repo.feeds']->find($id);
-        if (!$feed->isAccessible($app['authentication']->getUser(), $app)) {
-            $app->abort(404, 'Feed not found');
-        }
-        $link = $app['feed.user-link-generator']->generate(
-            $feed,
-            $app['authentication']->getUser(),
-            FeedLinkGenerator::FORMAT_RSS,
-            null,
-            $renew
-        );
-
-        $output = [
-            'texte' => '<p>' . $app->trans(
-                    'publication::Voici votre fil RSS personnel. Il vous permettra d\'etre tenu au courrant des publications.'
-                ) . '</p><p>' . $app->trans('publications::Ne le partagez pas, il est strictement confidentiel') . '</p>
-            <div><input type="text" style="width:100%" value="' . $link->getURI() . '"/></div>',
-            'titre' => $app->trans('publications::votre rss personnel')
-        ];
-
-        return $app->json($output);
-    }
-
-    public function ensureUserHasPublishRight(Application $app)
-    {
-        $this->getFirewall($app)->requireRight('bas_chupub');
     }
 }
